# swagger.yaml
openapi: 3.0.0
info:
  title: Social Media Express API
  version: 1.0.0
  description: Projet Web-API M1 DFS
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
...
paths:
...
  /api/comment:
    get:
      summary: Get all comments
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of all comments
          content:
            application/json:
              schema:
                type: object
                properties:
                  comments:
                    type: array
                    items:
                      type: object
        '500':
          description: Server error
  /api/comment/getAllByUser/{id}:
    get:
      summary: Get comments by user ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of comments for the specified user
          content:
            application/json:
              schema:
                type: object
                properties:
                  comments:
                    type: array
                    items:
                      type: object
        '400':
          description: Bad request - user ID required
        '500':
          description: Server error
  /api/comment/create:
    post:
      summary: Create a new comment
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
                - userId
                - articleId
              properties:
                content:
                  type: string
                userId:
                  type: string
                articleId:
                  type: string
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Bad request - missing required fields
        '500':
          description: Server error
  /api/content:
    get:
      summary: Get all contents
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of all contents
          content:
            application/json:
              schema:
                type: object
                properties:
                  contents:
                    type: array
                    items:
                      type: object
        '500':
          description: Server error
...
  /api/content/create:
    post:
      summary: Create new content
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - userId
              properties:
                title:
                  type: string
                userId:
                  type: string
      responses:
        '201':
          description: Content created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Bad request - missing required fields
        '500':
          description: Server error
...